requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
langchain==0.1.20
chromadb==0.3.29
python-dotenv==1.0.0
pydantic==1.10.10
aiofiles==23.1.0
requests==2.31.0
